pyproject.toml
[project]
name = "Flask-Esports"
version = "0.1"
description = "A basic framework for rapidly developing esports APIs based around web scraping"
readme = "README.md"
license = { file = "LICENSE.txt" }
authors = [
    {name = "Jopat2409", email = "joantpat@gmail.com"}
]
maintainers = [
    {name = "Jopat2409", email = "joantpat@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "flask",
    "python-dotenv",
    "lxml",
    "requests"
]

[project.urls]
Source = "https://github.com/Jopat2409/esports-api"

[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "flask_esports"

[tool.flit.sdist]
include = [
    "esports_api/",
    "scripts/",
    "tests/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["flask_esports", "tests"]

[tool.ruff]
exclude = ["tests/*.py"] # Exclude tests from ruff linting

[tool.ruff.lint.per-file-ignores]
# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
"__init__.py" = ["E402"]
